requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
ollama==0.6.1
openai-whisper==20250625
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
regex==2026.1.15
requests==2.32.5
scipy==1.13.1
sounddevice==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.8.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
